+++ src/main/java/songbird/command/ByeCommand.java
@@ -1,4 +1,4 @@
-package songbird.commands;
+package songbird.command;
 
 import songbird.ui.Ui;
