+++ src/main/java/songbird/command/ListCommand.java
@@ -1,6 +1,6 @@
-package songbird.commands;
+package songbird.command;
 
-import songbird.tasks.TaskList;
+import songbird.task.TaskList;
 import songbird.ui.Ui;
 
 /**
